Remove window arch
